Finshed the valid move checker, and moved main into its own file.
This commit adds the rest of the valid move checking to the board. This
commit also refactors the main function out into its own file. This main
function only creates a new board, and prints it out to the screen, for
now. The pawn valid move checker is still a WIP.