Refactors the board for more explicit ownership of pieces.
This commit changes the board to store unique_ptr<const Piece> instead
of simply const Piece* for each piece on the board. This results in
cleaner constructors and destructors, in addition to more explicit
ownership transfer when moving a piece. Additionally, modify the print
behavior of the board to print out more useful information, and modify
the print behavior of pieces to print out Unicode chess symbols, instead
of ANSI colored letters.